+++ forge/changelog.md
@@ -11,3 +11,4 @@
 - Added customizable group HUD position
 - Added customizable HUD icon position
 - Fixed voice chat settings can't be opened when not connected
+- Added thread pooling to audio recording